## Description
Following d6c6351c7f0b241578121d459943fdf2b305571e, the `ir.rule` related to discuss channels got more elaborated, with more subqueries. When loading an `avatar_128`, in an override of `_find_record_check_access`, a repetitive domain with an `any` operator on the same left leaf. This repetitive domain + the new `ir.rule` lead to a bad postgres plan, leading to a perf regression.

## Fix
Aggregate the right leaf of the domain in a disjonction, it's semantically equivalent, and leads to less subqueries in total for the final query. The plan is simpler and actually uses lazy merging nodes, which allows early exit with the `limit=1`, instead of hashjoins, which forces a total evaluation of the subplans (so `limit=1` is useless in that case).

## Benchmark
On next.odoo.com, loading an `avatar_128`, cache disabled in browser (simulating a first load)

|               | Before | After | Speed-up |
|---------------|--------|-------|----------|
| Timings (hot) | 3s     | 200ms | 15x      |

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr