@robodoo 2024-11-13
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/187072.png)](https://mergebot.odoo.com/odoo/odoo/pull/187072)