@pimodoo 2024-11-13
@pivi-odoo It seems the runbot is red because of avatar stuff. So your PR probably break them (and not a false positive)