@pivi-odoo 2024-11-15
Nature of the patch changed, previous reviews should be considered out-dated.