@pivi-odoo 2024-11-15
Patch has been approved and effectiveness has been validated.
Merging to be available for next deployment.

@robodoo r+